moved commands and handlers to different files
